Split component and styles out to smaller scopes
