+++ src/components/cc-form/styles.module.scss
@@ -1,3 +1,6 @@
+@use './styles/input';
+@use './styles/submit';
+
 .form {
   display: grid;
   align-items: end;
@@ -16,7 +19,7 @@
     grid-template-columns: repeat(10, 1fr);
     padding: 16px;
 
-    > div.cc_number, > button.submitButton {
+    > div.cc_number, > button.submit {
       grid-column: span 10;
     }
     > fieldset.cc_exp {
@@ -27,39 +30,7 @@
     }
   }
 
-  label, legend {
-    display: inline-block;
-    padding: 0;
-    margin-bottom: 4px;
-
-    font-size: 1rem;
-  }
-
-  input {
-    color: inherit;
-    line-height: inherit;
-    background: inherit;
-    transition: all 0.3s ease;
-
-    box-sizing: border-box;
-    width: 100%;
-    height: 2.8rem;
-    padding: 0;
-    padding-inline: 1rem;
-
-    background-color: hsl(215, 50%, 94%);
-    border: 1px solid rgb(175, 180, 200);
-    border-radius: 4px;
-
-    color: rgba(0, 0, 0, 0.8);
-    font-size: 1.5rem;
-
-    &:focus {
-      outline: none;
-      border-color: rgb(0, 128, 255);
-      box-shadow: 0 0 0 4px dodgerblue;
-    }
-  }
+  @include input.label-legend();
 }
 
 
@@ -92,37 +63,5 @@ div.cc_csc {
       min-width: calc(4ch + 32px);
       text-align: center;
     }
-  }
-}
-
-
-button.submitButton {
-  display: block;
-  cursor: pointer;
-  transition: all 0.3s ease;
-
-  margin-top: auto;
-  padding: 8px 32px;
-
-  width: 100%;
-  height: 2.8rem;
-
-  background-color: rgb(47, 104, 179);
-  border: none;
-  border-radius: 4px;
-
-  color: white;
-  font-size: 1rem;
-  font-weight: bold;
-  text-align: center;
-
-  &:active, &:hover {
-    background-color: dodgerblue;
-  }
-
-  &:focus {
-    outline: none;
-    background-color: dodgerblue;
-    box-shadow: 0 0 0 4px dodgerblue;
   }
 }
