+++ src/components/cc-form/styles/_submit.scss
@@ -1,0 +1,30 @@
+button.submit {
+  display: block;
+  cursor: pointer;
+  transition: all 0.3s ease;
+
+  margin-top: auto;
+  padding: 8px 32px;
+
+  width: 100%;
+  height: 2.8rem;
+
+  background-color: rgb(47, 104, 179);
+  border: none;
+  border-radius: 4px;
+
+  color: white;
+  font-size: 1rem;
+  font-weight: bold;
+  text-align: center;
+
+  &:active, &:hover {
+    background-color: dodgerblue;
+  }
+
+  &:focus {
+    outline: none;
+    background-color: dodgerblue;
+    box-shadow: 0 0 0 4px dodgerblue;
+  }
+}
